pyproject.toml
[tool.poetry]
name = "mantraglyph"
version = "0.0.1"
description = ""
authors = ["AIEAJN <nathanuchia@gmail.com>"]
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
pandas = "^1.2.4"
opencv-python = "^4.5.1"
pytesseract = "^0.3.7"
translators = "^4.7.16"
google-cloud-translate = "^3.11.3"
matplotlib = "^3.7.2"
manga-ocr = "^0.1.10"
django = "^5.0.6"
graphene-django = "^3.1.5"
graphene-luna = "^1.0.0"
psycopg2 = "^2.9.9"
google-cloud-vision = "^3.4.5"
django-filter = "^23.5"
django-cors-headers = "^4.3.1"
django-allow-cidr = "^0.7.1"
graphene-file-upload = "^1.3.0"
google-cloud-storage = "^2.13.0"
firebase = "^4.0.1"
firebase-admin = "^6.3.0"
uvicorn = "^0.24.0.post1"
django-graphql-jwt = "^0.4.0"
whitenoise = "^6.6.0"
gunicorn = "^21.2.0"
graphene-django-query-optimizer = "^0.8.4"
psycopg = {extras = ["binary"], version = "^3.2.1"}

[tool.poetry.dev-dependencies]
pytest = "^5.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
